+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 - Impl `Hash`, `PartialEq`, and `Eq` for `events::ModifiersState`.
+- Implement `MonitorId::get_hidpi_factor` for MacOS.
+- Added method `os::macos::MonitorIdExt::get_nsscreen() -> *mut c_void` that gets a `NSScreen` object matching the monitor ID.
 
 # Version 0.11.1 (2018-02-19)